requirements.txt
# Core dependencies
pynput>=1.7.6
cryptography>=41.0.0
requests>=2.31.0
Pillow>=10.0.0
configparser>=5.3.0
psutil>=5.9.0

# Windows-specific dependencies
pywin32>=306; sys_platform == "win32"
pywin32-ctypes>=0.2.2; sys_platform == "win32"

# Scheduling and utilities
schedule>=1.2.0
python-dateutil>=2.8.2

# Optional dependencies for enhanced functionality
colorama>=0.4.6
tqdm>=4.66.0

# Development and testing dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

